Remove temp.elf.elf and temp.hex.hex at the beginning of the avr-make script
Currently, if temp.elf.elf exists before the script is run, the script will always believe avr-gcc has run successfully regardless of whether or not it failed (and likewise for temp.hex.hex + avr-objcopy). Removing the files at
the beginning of script should ensure that the script only proceeds if avr-gcc actually succeeded.